+++ src/interfaces.ts
@@ -15,13 +15,6 @@ export interface IUsers extends Document {
     complement: string
     [key: string]: any
 }
-export interface IUserDb {
-    _id: string;
-    name: string;
-    lastName: string;
-    email: string;
-  }
-  
 
 export interface IEntityDb {
     _id: string;
